Aggregate view permissions for `Patch` to `cluster-reader`
We only aggregate `view` permissions for `Patch` to `cluster-reader`,
since unprivileged users can't really use the patch-operator anyway,
since the operator needs cluster-level list for any resource targeted by
a patch, which regular users won't be able to grant to their service
accounts.

By not giving access to the `Patch` resource throught the regular
`view`, `edit` and `admin` cluster roles we avoid that users discover
the resource and start trying to use it which would only generate errors
in the operator.