Fix `vault kv put` command in external certificates how-to
We need to either use `put` followed by `patch` to end up with a Vault
secret which contains both private keys, or specify both keys in a
single `vault put`.

This commit opts for the latter variant, and creates the secret with
both fields in a single `vault kv put` command.